requirements.txt
Flask
gunicorn
google-cloud
google-cloud-firestore
git+git://github.com/keenetic/cloud-api-python-client@master#egg=ndcloudclient